refactor(characters): move design update functions to own service
